Remove recursive rule variable reference
The command variable was referring to the depfile variable in the same
rule scope. These inter-scope references are only possible in rules,
not builds or the global scope. (though the global scope can reference
earlier variables)

n2 doesn't want to support this because it would slow down evaluating
the build file, so remove the usage.

Bug: 318434287
Test: Presubmits
Change-Id: Iffbfe74dc7a400ff2fa827ab201995eeeba29280